+++ documentdb_tests/compatibility/tests/system/administration/commands/setUserWriteBlockMode/test_setUserWriteBlockMode_write_block_enforcement.py
@@ -1,7 +1,6 @@
-"""Tests for setUserWriteBlockMode write block enforcement.
+"""Tests for setUserWriteBlockMode write block enforcement errors.
 
-Validates that write operations are blocked while the block is active,
-read operations are not affected, and operations succeed when block is disabled.
+Validates that write operations are rejected while the block is active.
 """
 
 from __future__ import annotations
@@ -18,10 +17,7 @@ from documentdb_tests.compatibility.tests.system.administration.commands.setUser
 from documentdb_tests.compatibility.tests.system.administration.commands.utils.admin_test_case import (  # noqa: E501
     AdminTestCase,
 )
-from documentdb_tests.framework.assertions import (
-    assertFailureCode,
-    assertSuccessPartial,
-)
+from documentdb_tests.framework.assertions import assertFailureCode
 from documentdb_tests.framework.error_codes import USER_WRITES_BLOCKED_ERROR
 from documentdb_tests.framework.executor import execute_admin_command, execute_command
 from documentdb_tests.framework.parametrize import pytest_params
@@ -187,108 +183,8 @@ WRITE_BLOCKED_TESTS: list[AdminTestCase] = [
     ),
 ]
 
-# Property [Read Operations Not Blocked]: read operations succeed while the block is active.
-READ_NOT_BLOCKED_TESTS: list[AdminTestCase] = [
-    AdminTestCase(
-        "read_find",
-        use_admin=False,
-        partial_success=True,
-        docs=[{"_id": "read_doc", "x": 1}],
-        pre_command=lambda c: execute_admin_command(
-            c, {"setUserWriteBlockMode": 1, "global": True}
-        ),
-        command=lambda ctx: {"find": ctx.collection, "filter": {}},
-        expected={"ok": 1.0},
-        msg="setUserWriteBlockMode should not block find while active",
-    ),
-    AdminTestCase(
-        "read_aggregate",
-        use_admin=False,
-        partial_success=True,
-        docs=[{"_id": "read_doc", "x": 1}],
-        pre_command=lambda c: execute_admin_command(
-            c, {"setUserWriteBlockMode": 1, "global": True}
-        ),
-        command=lambda ctx: {
-            "aggregate": ctx.collection,
-            "pipeline": [{"$match": {}}],
-            "cursor": {},
-        },
-        expected={"ok": 1.0},
-        msg="setUserWriteBlockMode should not block aggregate while active",
-    ),
-    AdminTestCase(
-        "read_count",
-        use_admin=False,
-        partial_success=True,
-        docs=[{"_id": "read_doc", "x": 1}],
-        pre_command=lambda c: execute_admin_command(
-            c, {"setUserWriteBlockMode": 1, "global": True}
-        ),
-        command=lambda ctx: {"count": ctx.collection},
-        expected={"ok": 1.0},
-        msg="setUserWriteBlockMode should not block count while active",
-    ),
-    AdminTestCase(
-        "read_distinct",
-        use_admin=False,
-        partial_success=True,
-        docs=[{"_id": "read_doc", "x": 1}],
-        pre_command=lambda c: execute_admin_command(
-            c, {"setUserWriteBlockMode": 1, "global": True}
-        ),
-        command=lambda ctx: {"distinct": ctx.collection, "key": "x"},
-        expected={"ok": 1.0},
-        msg="setUserWriteBlockMode should not block distinct while active",
-    ),
-]
 
-# Property [Writes Succeed When Disabled]: write operations succeed when no block is active.
-WRITE_SUCCEEDS_TESTS: list[AdminTestCase] = [
-    AdminTestCase(
-        "succeeds_insert",
-        use_admin=False,
-        partial_success=True,
-        docs=[{"_id": "upd", "x": 1}, {"_id": "del"}],
-        command=lambda ctx: {"insert": ctx.collection, "documents": [{"_id": "ok"}]},
-        expected={"ok": 1.0},
-        msg="setUserWriteBlockMode should allow insert when block is not active",
-    ),
-    AdminTestCase(
-        "succeeds_update",
-        use_admin=False,
-        partial_success=True,
-        docs=[{"_id": "upd", "x": 1}, {"_id": "del"}],
-        command=lambda ctx: {
-            "update": ctx.collection,
-            "updates": [{"q": {"_id": "upd"}, "u": {"$set": {"x": 2}}}],
-        },
-        expected={"ok": 1.0},
-        msg="setUserWriteBlockMode should allow update when block is not active",
-    ),
-    AdminTestCase(
-        "succeeds_delete",
-        use_admin=False,
-        partial_success=True,
-        docs=[{"_id": "upd", "x": 1}, {"_id": "del"}],
-        command=lambda ctx: {
-            "delete": ctx.collection,
-            "deletes": [{"q": {"_id": "del"}, "limit": 1}],
-        },
-        expected={"ok": 1.0},
-        msg="setUserWriteBlockMode should allow delete when block is not active",
-    ),
-]
-
-ENFORCEMENT_TESTS: list[AdminTestCase] = (
-    WRITE_BLOCKED_TESTS + READ_NOT_BLOCKED_TESTS + WRITE_SUCCEEDS_TESTS
-)
-
-ENFORCEMENT_ERROR_TESTS: list[AdminTestCase] = WRITE_BLOCKED_TESTS
-ENFORCEMENT_SUCCESS_TESTS: list[AdminTestCase] = READ_NOT_BLOCKED_TESTS + WRITE_SUCCEEDS_TESTS
-
-
-@pytest.mark.parametrize("test", pytest_params(ENFORCEMENT_ERROR_TESTS))
+@pytest.mark.parametrize("test", pytest_params(WRITE_BLOCKED_TESTS))
 def test_setUserWriteBlockMode_blocked(database_client, collection, test):
     """Test setUserWriteBlockMode blocks write operations while active."""
     collection = test.prepare(database_client, collection)
@@ -296,13 +192,3 @@ def test_setUserWriteBlockMode_blocked(database_client, collection, test):
     ctx = CommandContext.from_collection(collection)
     result = execute_command(collection, test.build_command(ctx))
     assertFailureCode(result, test.error_code, msg=test.msg)
-
-
-@pytest.mark.parametrize("test", pytest_params(ENFORCEMENT_SUCCESS_TESTS))
-def test_setUserWriteBlockMode_allowed(database_client, collection, test):
-    """Test setUserWriteBlockMode allows reads and writes when appropriate."""
-    collection = test.prepare(database_client, collection)
-    test.run_pre_command(collection)
-    ctx = CommandContext.from_collection(collection)
-    result = execute_command(collection, test.build_command(ctx))
-    assertSuccessPartial(result, test.expected, msg=test.msg)
